Added testcases

+++ src/pke/examples/sigmoid-ckks.cpp
@@ -46,7 +46,7 @@ uint32_t multDepth = 4;
 template <typename T>
 auto mse(std::vector<double> original, std::vector<T> approx) {
     double error = 0;
-    for(int i = 0; i < original.size(); i++){
+    for(size_t i = 0; i < original.size(); i++){
         auto diff = 0;
         if constexpr (std::is_same<T, std::complex<double>>::value)
             diff = original[i] - approx[i].real();
@@ -62,7 +62,7 @@ auto mse(std::vector<double> original, std::vector<T> approx) {
 template <typename T>
 auto mae(std::vector<double> original, std::vector<T> approx) {
     double error = 0;
-    for(int i = 0; i < original.size(); i++){
+    for(size_t i = 0; i < original.size(); i++){
         auto diff = 0;
         if constexpr (std::is_same<T, std::complex<double>>::value)
             diff = original[i] - approx[i].real();
@@ -77,7 +77,7 @@ auto mae(std::vector<double> original, std::vector<T> approx) {
 template <typename T>
 auto mape(std::vector<double> original, std::vector<T> approx) {
     double error = 0;
-    for(int i = 0; i < original.size(); i++){
+    for(size_t i = 0; i < original.size(); i++){
         if(original[i] != 0) {
             double diff = 0;
             if constexpr (std::is_same<T, std::complex<double>>::value)
